Initial powerpc support
Below is a basic powerpc port of Zephyr. It runs on both qemu and
microwatt [1] and runs the basic tests and samples in the
repository. It supports both cooperative and pre-emptive multitasking.

The patch still contains a few FIXMEs but reviews are welcome. When we
are a bit more comfortable with the patch, we'll create a PR.

To get started do:
  export ZEPHYR_TOOLCHAIN_VARIANT=cross-compile
  export CROSS_COMPILE=/usr/local/powerpc64le-power8--glibc--bleeding-edge-2018.11-1/bin/powerpc64le-linux-
  export ZEPHYR_BASE=/path/to/zephyr/
  cmake -B build -GNinja -DBOARD=microwatt samples/hello_world
  cd build
  ninja
  qemu-system-ppc64 -M powernv -cpu POWER9 -nographic -bios zephyr/zephyr.bin

Port done by Anton Blanchard and Michael Neuling.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by Anton Blanchard <anton@linux.ibm.com>